app/requirements.txt
cherrypy
Flask
pyyaml
